Remove AWS dependant import/test from cwlTest.py.
